+++ composeApp/src/commonMain/kotlin/com/example/mobile_client_app/features/classes/bookingClass/domain/usecase/GetClassDetailsUseCase.kt
@@ -9,7 +9,7 @@ class GetClassDetailsUseCase(
     private val repository: ClassRepository
 ) {
     suspend operator fun invoke(
-        classId: Int,
+        classId: Long,
     ): Result<ClassDetailsResponse, NetworkError> {
         return repository.getClassDetails(
             classId = classId,
