+++ composeApp/src/commonMain/kotlin/com/example/mobile_client_app/features/classes/bookingClass/domain/usecase/SendBookingClassRequestUseCase.kt
@@ -8,7 +8,7 @@ class SendBookingClassRequestUseCase(
     private val repository: ClassRepository
 ) {
     suspend operator fun invoke(
-        classId: Int
+        classId: Long
     ): Result<Unit, NetworkError> {
         return repository.sendBookingClassRequest(classId)
     }
