+++ composeApp/src/commonMain/kotlin/com/example/mobile_client_app/features/classes/bookingClass/presentation/BookingClassScreen.kt
@@ -38,7 +38,7 @@ import org.koin.compose.viewmodel.koinViewModel
 @Composable
 fun BookingClassScreen(
     viewModel: BookingClassViewModel = koinViewModel(),
-    classId: Int,
+    classId: Long,
     onCancelBookingClick: () -> Unit,
     onConfirmClick:() -> Unit,
 ) {
