+++ composeApp/src/commonMain/kotlin/com/example/mobile_client_app/features/onboarding/classes/data/remote/ClassesApi.kt
@@ -7,6 +7,6 @@ import com.example.mobile_client_app.util.network.Result
 
 interface ClassesApi {
     suspend fun getClasses(): Result<List<FitnessClassResponse>, NetworkError>
-    suspend fun getClassDetails(classId: Int): Result<ClassDetailsResponse, NetworkError>
-    suspend fun sendBookingClassRequest(classId: Int): Result<Unit, NetworkError>
+    suspend fun getClassDetails(classId: Long): Result<ClassDetailsResponse, NetworkError>
+    suspend fun sendBookingClassRequest(classId: Long): Result<Unit, NetworkError>
 }
